(new incomplete major-api) Working towards a more asynchronicity-friendly API.
This is just a start; we’re ensuring each `CEST()` is “aware” (by way of a local variable) of “itself” (the `struct cest` on the heap representing the test being run.) This will later allow us to make a `Cest.complete()` call with a reference to “our” `struct cest`, which will allow us to eschew return-values-as-state, and instead store result state in the `cest` itself.

After that, it’s just a matter of cosmetics (printing completed `CEST()`s in order, even if they finish out-of-order).