(new) Adding support for a `SOMEDAY;` directive.
This also involved slightly modifying the semantics (but not functionality) of `PEND;`. The old `PEND;` has been split into the current `SOMEDAY;` and `PEND;` directives. `SOMEDAY;`s are intended to be used before tests that have been (at least partially) written, but whose tested-functionality is not actually implemented at the moment, and will not be for a while. `PEND;` is for functionality being implemented *now*; a `PEND;` is considered a failure, but is visually differentiable from a `FAIL;`.

Use the new `PEND;` when you wish to “comment out” the latter portions of a test while you’re modifying the implementation of a feature; this allows you to differentiate failure of a particular portion of the test (the portion you’re working on implementing) and failure in later parts, which you won’t be able to make pass until you’ve finished with the current portion.

When your suite completes *with pending tests* (magenta output, currently), it implies you’re not ready to commit, but all pasts are technically *passing* (no ASSERTions that were actually executed *failed*, but some ASSERTions are still “commented out” with `PEND;`).